Commit 7

+++ MovieRentalApp/Rental.java
@@ -37,8 +37,8 @@ public class Rental {
         
         return amount;
    }
-   public double getFrequentRenterPoints(){
-   	double frequentRenterPoints = 0;
+   public int getFrequentRenterPoints(){
+   	int frequentRenterPoints = 0;
    	// add frequent renter points
         frequentRenterPoints ++;
         // add bonus for a two day new release rental
